minor things

+++ source/measures.cpp
@@ -84,13 +84,13 @@ void energy(struct Measures &mis, struct Villain &vil, const std::vector<Node> &
                 i = ix + Lx * (iy + iz * Ly);
                 for(vec=0; vec<3; vec++) {
                     if(vec==0){
-                        nn_i= mod(ix+1, (int)Lx) + Lx * (iy + iz * Ly);
+                        nn_i= mod(ix+1, Lx) + Lx * (iy + iz * Ly);
                     }
                     if(vec==1){
-                        nn_i= ix + Lx * (mod(iy+1,(int)Ly) + iz * Ly);
+                        nn_i= ix + Lx * (mod(iy+1,Ly) + iz * Ly);
                     }
                     if(vec==2){
-                        nn_i= ix + Lx * (iy + mod(iz+1,(int)Lz) * Ly);
+                        nn_i= ix + Lx * (iy + mod(iz+1,Lz) * Ly);
                     }
 
                     arg_1 = arg((Site[nn_i].Psi[0] - Site.at(i).Psi[0]), MaxP);
@@ -175,7 +175,6 @@ void helicity_modulus(struct Measures &mis, struct Villain &vil, const std::vect
 //}
 
 void magnetization_singlephase(struct Measures &mis, const std::vector<Node> &Site){
-    //The Ising parameter m(x,y)=+/-1 indicates the chirality of the three phases. If the phases are ordered as: phi_1, phi_2, phi_3 then m=1; otherwise if the order is phi_1, phi_3, phi_2 then m=-1.
     unsigned alpha;
     double cos_phi[NC]={0}, sin_phi[NC]={0};
     double inv_N=1./N;
